Add server repo and service
