+++ src/services/event.serverService.ts
@@ -1,0 +1,50 @@
+import { BaseEvent, NewEvent } from "@/model/event";
+import { addEvent, deleteEventById } from "@/repositories/event.serverRepository";
+
+export type EventsData = {
+    events: BaseEvent[],
+    firstPage: number, 
+    pageSize: number,
+    totalCount: number,
+    index: number
+}
+
+export const eventServerService = {
+        async addUpcomingEvent(event: NewEvent) {
+            try {
+                if (!event.name) {
+                    throw new Error("Event name is required");
+                }
+    
+                if (event.image_path && event.image_file) {
+                    await uploadImage(event.image_path, event.image_file);
+                }
+    
+                await addEvent(event);
+            } catch (error) {
+                throw new Error("eventService.addUpcomingEvent failed", { cause : error });
+            }
+        },
+    
+        async deleteEvent(event: BaseEvent) {
+            try {
+                await deleteEventById(event.id);
+                if (event.image_path) {
+                    await deleteImage(event.image_path);
+                }
+            } catch (error) {
+                throw new Error("eventService.deleteEvent failed", { cause : error })
+            }
+        },
+
+        async updateEventInfo(event: BaseEvent) {
+            try {
+                if (event.image_file && event.image_path) {
+                    await updateImage(event.image_path, event.image_file);
+                }
+                await updateEvent(event);
+            } catch (error) {
+                throw new Error("eventService.updateEventInfo failed", { cause : error })
+            }
+        },
+}
